+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.6.0 (unreleased)
 ------------------
 
+- #2562 Increased field size and allowed input length for instrument import folder path
 - #2555 Instrument Results Importer/Parser improvements and fixtures
 - #2551 Migrate InstrumentTypes to Dexterity
 - #2560 Fix no profiles in add form on sample type and/or client change